Feat : 판매중, 판매완료 api 연결 확인
