+++ src/api/request.tsx
@@ -140,6 +140,7 @@ export const getClothingSalesDetails = async (
         method: "GET",
         headers: {
           "Content-Type": "application/json",
+          Authorization: `Bearer ${process.env.API_TOKEN}`,
         },
       }
     );
